sheta

+++ Client/Models/Spr_oplat_sklad.cs
@@ -5,7 +5,7 @@ namespace Client.Models
     public class Spr_oplat_sklad
     {
         [Key]
-        public decimal kod_zap { get; set; }
+        public int kod_zap { get; set; }
         public string naim { get; set; }
     }
 }
